Move LintCommand logic into LintOrAnalyzeCommand
to prepare for the upcoming analyze command
