+++ Source/swiftlint/Commands/LintCommand.swift
@@ -1,118 +1,12 @@
 import Commandant
-import Dispatch
-import Foundation
 import Result
-import SourceKittenFramework
-import SwiftLintFramework
 
 struct LintCommand: CommandProtocol {
     let verb = "lint"
     let function = "Print lint warnings and errors (default command)"
 
     func run(_ options: LintOptions) -> Result<(), CommandantError<()>> {
-        var fileBenchmark = Benchmark(name: "files")
-        var ruleBenchmark = Benchmark(name: "rules")
-        var violations = [StyleViolation]()
-        let configuration = Configuration(options: options)
-        let reporter = reporterFrom(options: options, configuration: configuration)
-        let cache = options.ignoreCache ? nil : LinterCache(configuration: configuration)
-        let visitorMutationQueue = DispatchQueue(label: "io.realm.swiftlint.lintVisitorMutation")
-        return configuration.visitLintableFiles(options: options, cache: cache) { linter in
-            let currentViolations: [StyleViolation]
-            if options.benchmark {
-                let start = Date()
-                let (violationsBeforeLeniency, currentRuleTimes) = linter.styleViolationsAndRuleTimes
-                currentViolations = LintCommand.applyLeniency(options: options, violations: violationsBeforeLeniency)
-                visitorMutationQueue.sync {
-                    fileBenchmark.record(file: linter.file, from: start)
-                    currentRuleTimes.forEach { ruleBenchmark.record(id: $0, time: $1) }
-                    violations += currentViolations
-                }
-            } else {
-                currentViolations = LintCommand.applyLeniency(options: options, violations: linter.styleViolations)
-                visitorMutationQueue.sync {
-                    violations += currentViolations
-                }
-            }
-            linter.file.invalidateCache()
-            reporter.report(violations: currentViolations, realtimeCondition: true)
-        }.flatMap { files in
-            if LintCommand.isWarningThresholdBroken(configuration: configuration, violations: violations)
-                && !options.lenient {
-                violations.append(LintCommand.createThresholdViolation(threshold: configuration.warningThreshold!))
-                reporter.report(violations: [violations.last!], realtimeCondition: true)
-            }
-            reporter.report(violations: violations, realtimeCondition: false)
-            let numberOfSeriousViolations = violations.filter({ $0.severity == .error }).count
-            if !options.quiet {
-                LintCommand.printStatus(violations: violations, files: files,
-                                        serious: numberOfSeriousViolations)
-            }
-            if options.benchmark {
-                fileBenchmark.save()
-                ruleBenchmark.save()
-            }
-            try? cache?.save()
-            return LintCommand.successOrExit(numberOfSeriousViolations: numberOfSeriousViolations,
-                                             strictWithViolations: options.strict && !violations.isEmpty)
-        }
-    }
-
-    private static func successOrExit(numberOfSeriousViolations: Int,
-                                      strictWithViolations: Bool) -> Result<(), CommandantError<()>> {
-        if numberOfSeriousViolations > 0 {
-            exit(2)
-        } else if strictWithViolations {
-            exit(3)
-        }
-        return .success(())
-    }
-
-    private static func printStatus(violations: [StyleViolation], files: [File], serious: Int) {
-        let pluralSuffix = { (collection: [Any]) -> String in
-            return collection.count != 1 ? "s" : ""
-        }
-        queuedPrintError(
-            "Done linting! Found \(violations.count) violation\(pluralSuffix(violations)), " +
-            "\(serious) serious in \(files.count) file\(pluralSuffix(files))."
-        )
-    }
-
-    private static func isWarningThresholdBroken(configuration: Configuration,
-                                                 violations: [StyleViolation]) -> Bool {
-        guard let warningThreshold = configuration.warningThreshold else { return false }
-        let numberOfWarningViolations = violations.filter({ $0.severity == .warning }).count
-        return numberOfWarningViolations >= warningThreshold
-    }
-
-    private static func createThresholdViolation(threshold: Int) -> StyleViolation {
-        let description = RuleDescription(
-            identifier: "warning_threshold",
-            name: "Warning Threshold",
-            description: "Number of warnings thrown is above the threshold.",
-            kind: .lint
-        )
-        return StyleViolation(
-            ruleDescription: description,
-            severity: .error,
-            location: Location(file: "", line: 0, character: 0),
-            reason: "Number of warnings exceeded threshold of \(threshold).")
-    }
-
-    private static func applyLeniency(options: LintOptions, violations: [StyleViolation]) -> [StyleViolation] {
-        if !options.lenient {
-            return violations
-        }
-        return violations.map {
-            if $0.severity == .error {
-                return StyleViolation(ruleDescription: $0.ruleDescription,
-                                      severity: .warning,
-                                      location: $0.location,
-                                      reason: $0.reason)
-            } else {
-                return $0
-            }
-        }
+        return LintOrAnalyzeCommand.run(LintOrAnalyzeOptions(options))
     }
 }
